% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annualTrends}
\alias{annualTrends}
\title{Annual Trends}
\usage{
annualTrends(Stations = "All", Snames = "All", use.logs = FALSE,
  ar.adj = TRUE, report)
}
\arguments{
\item{Stations}{a vector of the the station identifiers on which
to do the flow adjustment.}

\item{Snames}{a vector of the response variables on which to do
the flow adjustment.}

\item{use.logs}{log transform the data before the trend test?}

\item{ar.adj}{adjust the attained p-value to account for serial
correlation? See \bold{Details}.}

\item{report}{the name of the PDF file that contains a report for
each test. The default is to use the name of the project with "_an"
appended. If the PDF file exists, then it is not overwritten,
but the name is appended with a sequence of numbers until one that
is valid is created.}
}
\value{
The name of the report file.
}
\description{
Perform the annual Kendall trend test.
}
\details{
The \code{kensen.test} includes a correction to the attained p-value
that accounts the first order auto regression of the data. It is
applied whenever there are at least 10 observations. To suppress the
correction, set \code{ar.adj} to \code{FALSE}, or a numeric value can
be specified for the minimum number of observations, whihc can be
useful for a ragged analysis. The value for \code{ar.adj} cannot be
set to less than 10.
}

