% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ckProj.R
\name{ckProj}
\alias{ckProj}
\title{Check the ESTREND Project}
\usage{
ckProj()
}
\value{
The position of the current project in the serach path
is returned.
}
\description{
Check and get the ESTREND project to use for this analyses.
}

