% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estrendMonthTable.R
\name{estrendMonthTable}
\alias{estrendMonthTable}
\title{Evaluate Monthly Definitions}
\usage{
estrendMonthTable(ssn12, frctn = 0.5)
}
\arguments{
\item{ssn12}{the vector of data selected for the monthly
seaken analysis.}

\item{frctn}{limit for the fraction of possible observations in
each half.}
}
\value{
A list of the results for the percentage of observations in each
half, the selected months, and the number of selected months.
}
\description{
Create tables of observations for each season in both the first and last
halfs of the record. Internal use only.
}

