% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estrendSeasonTable}
\alias{estrendSeasonTable}
\title{Evaluate Seasonal Definitions}
\usage{
estrendSeasonTable(ssn12, frctn = 0.5, pctg = 0.8)
}
\arguments{
\item{ssn12}{the vector of data selected for the 12 seasons per
year seaken analysis.}

\item{frctn}{limit for the fraction of possible observations in
the BE.}

\item{pctg}{required percentage of seasons needed to exceed frctn.}
}
\value{
A list of the results for 12, 6, 4, and 3 seasons per year
and the "best."
}
\description{
Create tables of observations for each season in both the BE and MI
records. Internal use only.
}

