% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTrends}
\alias{getTrends}
\title{Get Trends}
\usage{
getTrends(Stations = "All", Snames = "All", sig.level = 0.05)
}
\arguments{
\item{Stations}{a vector of the station identifiers on which
to do the flow adjustment.}

\item{Snames}{a vector of the response variables on which to do
the flow adjustment.}

\item{sig.level}{the alpha level of the test. If the attained p-value
of the test is less than \code{sig.level}, then the null hypothesis of
no trend is rejected and the trend is classes as "up" or "down."}
}
\value{
A data frame of the requested trend tests having these columns:
\item{Station}{ the station identifier.}
\item{Response}{ the response variable name.}
\item{Type}{ the type of trend test.}
\item{NumYears}{ the number of years for the trend test.}
\item{NumSeasons}{ the number of seasons. Set to \code{NA} for the Tobit
trend test.}
\item{Nobs}{ the number of observations used in the test.}
\item{RepValue}{ the representative value of the response variable. See
\code{Note}.}
\item{Trend}{ the trend expressed as the average rate in Response units
per year}
\item{Trend.pct}{ the trend expressed as the average rate in percent
change per year}
\item{P.value}{ the attained p-value of the test.}
\item{Trend.dir}{ an indicator of the trend. \code{Trend.dir} is "none" if the
attained p-value is greater than \code{sig.level}, "up" if the
attained p-value is less than \code{sig.level} and \code{Trend} is positive,
and "down" if the attained p-value is less than \code{sig.level} and
\code{Trend} is negative. For the seasonal Kendall test, when there are many
tied values, the results from the trend test and the trend may not agree,
in these cases \code{Trend.dir} will be "*."}
}
\description{
Extract the trend results.
}
\note{
The representative value is the median from a sample of the data
taken to represent a reasonably uniform sampling over time and throughout
the year. For the seasonal Kendall test, it is the median of the data, for
the Tobit test, it is the median of the data sampled as for a seasonal
Kendall test.
}

