% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/restrend-package.R
\docType{package}
\name{restrend-package}
\alias{restrend}
\alias{restrend-package}
\title{Estimate Trends (ESTREND)}
\description{
This package has specialized functions for managing data to
facilitate testing for linear or monotonic trends in hydrologic data.
}
\details{
\tabular{ll}{ Package: \tab restrend\cr
Type: \tab Package\cr
Version: \tab 0.4.1\cr
Date: \tab 2015-12-04\cr
License: \tab File LICENSE\cr
Depends: \tab g.data, smwrBase, smwrGraphs, smwrStats, smwrQW\cr }
This package contains functions that facilitate testing for linear or monotonic trends
in hydrologic data. Water-quality data or any other data collected on a
nearly regular basis can be uncensored, left censored, or multiply censored.
Data for annual analysis must be uncensored.
}
\author{
Dave Lorenz <lorenz@usgs.gov>

Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
Lorenz, D.L., in preparation, restrend---an R package for trend
estimation in hydrologic data, version 0.4.1
}
\seealso{
\code{\link[smwrQW:smwrQW-package]{smwrQW}}
}
\keyword{package}

