% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RSKTrends.R
\name{RSKTrends}
\alias{RSKTrends}
\title{Regional Seasonal Kendall Trends}
\usage{
RSKTrends(Stations = "All", Sname, FAC = FALSE)
}
\arguments{
\item{Stations}{a vector of the the station identifiers on which
to do the trend test.}

\item{Sname}{the response variables on which to do
the trend test.}

\item{FAC}{logical, if \code{TRUE}, then do the regional test on the
flow ajusted values, otherwise the analysis in done on the raw values.}
}
\value{
An object of class "rsktest" from \code{regionalSeaken}.
}
\description{
Perform the regional seasonal Kendall trend test. The test is only
appropriate for uncensored data and is performed after the trends have
been computed.
}
\seealso{
\code{\link{regionalSeaken}}
}

