% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SKTrends.R
\name{SKTrends}
\alias{SKTrends}
\title{Seasonal Kendall Trends}
\usage{
SKTrends(Stations = "All", Snames = "All", use.logs = TRUE,
  max.cens = 5, nseas = NULL, report)
}
\arguments{
\item{Stations}{a vector of the the station identifiers on which
to do the trend test.}

\item{Snames}{a vector of the response variables on which to do
the trend test.}

\item{use.logs}{logical, if \code{TRUE}, then log transform the data
before the trend test, otherwise no log transform is used. Applies only
to uncensored seasonal Kendall test---the data for the censored seasonal
Kendall test is never log-transformed.}

\item{max.cens}{the maximum percent censoring permitted for the
uncensored seasonal Kendall test. If the percentage of censoring
exceeds this value, then the censored seasonal Kendall test is
performed. Set to a negative value to force the censored seasonal
Kendall test for all \code{Stations} and \code{Snames}.}

\item{nseas}{the number of seasons to use for all of the tests. If
\code{NULL} (default), then use the selected number of seasons
defined in \code{setProj}. Applies only when the type of analysis
is "seasonal."}

\item{report}{the base name of the PDF file that contains a report for
each test; the suffix ".pdf" should not be inlcuded. The default is to
use the name of the project with "_sk" appended. If the PDF file exists,
then it is not overwritten, but the name is appended with a sequence of
numbers until one that is valid is created.}
}
\value{
The name of the report file.
}
\description{
Perform the seasonal Kendall trend test.
}

