% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tobitTrends}
\alias{tobitTrends}
\title{Tobit Trends}
\usage{
tobitTrends(Stations = "All", Snames = "All", use.logs = TRUE,
  Flow = TRUE, Seasons = TRUE, Covars = NULL, report)
}
\arguments{
\item{Stations}{a vector of the station identifiers on which
to do the flow adjustment.}

\item{Snames}{a vector of the response variables on which to do
the flow adjustment.}

\item{use.logs}{log transform the data (and flow) before the trend test?}

\item{Flow}{logical, include the flow variable? Can also be an expression
that includes any transformation of flow.}

\item{Seasons}{include first-order fourier seasonal terms? Can also be
an integer value indicating the order of the terms to include.}

\item{Covars}{the covariates listed in the regression model.}

\item{report}{the name of the PDF file that contains a report for
each test. The default is to use the name of the project with "_tb"
appended. If the PDF file exists, then it is not overwritten,
but the name is appended with a sequence of numbers until one that
is valid is created.}
}
\value{
The name of the report file.
}
\description{
Perform the tobit regression trend test.
}
\note{
The trend variable name is \code{.Dectime}, which represents
the trend in units of per year.
}

