\name{useProj}
\alias{useProj}
\title{Define the ESTREND Project}
\usage{
  useProj(project)
}
\arguments{
  \item{project}{the name of the project to use. The
  project name is forced to all lower case. If missing,
  then the most recent project is opened for use.}
}
\value{
  The name of the project is returned.
}
\description{
  Define the ESTREND project to use for subsequent
  analyses.
}

