% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regionalSeaken.R
\name{regionalSeaken}
\alias{regionalSeaken}
\title{Trend Test}
\usage{
regionalSeaken(series, nseas = 12)
}
\arguments{
\item{series}{a matrix of the regular series for each site. Each
column is the regular series of observations for that site.
Missing values are permitted.}

\item{nseas}{the number of seasons per year.}
}
\value{
An object of class "rsktest" containing the following components:
 \item{method}{a description of the method.}
 \item{statistic}{the value of Kendall's tau.}
 \item{p.values}{the p-value. See \bold{Note}.}
 \item{estimate}{the Sen estimate of the
slope in units per year--the median of the site medians.}
 \item{data.name}{a string containing the actual name of the input
series with the number of years and seasons.}
}
\description{
Compute the regional seasonal Kendall trend test.
}
\note{
The values of \code{p.values} are the attained p-values considering
the raw results, corrected for serial correlation, corrected for spatial
correlation using the method of Dietz and Killeen (1981), and the corrected
value for spatial correlation using the method of Douglas and others (2000).
}
\examples{
\dontrun{
data(RK3b)
# Build matrix, site 11 is missing 2004, would be row 60
Ammonia <- with(RK3b, matrix(c(value[1:59], NA, value[60:299]),
 ncol=25))
regionalSeaken(Ammonia, 1)
}
}
\references{
Dietz, E.J., and Killeen, T.J., 1981, A nonparametric multivariate test for
monotone trend with pharmaceutical applications: Journal of the American
Statistical Association, v. 76, p 169--174.\cr
Douglas, E.M., Vogel, R.M., and Kroll, C.N., 2000, Trends in floods and low
flows in the United States: impact of spatial correlation: Journal of
Hydrology, v. 240, p. 90--105.
Sprague, L.A., Mueller, D.K., Schwarz, G.E., and Lorenz, D.L., 2009,
Nutrient trends in streams and rivers of the United States, 1993--2003:
U.S. Geolgical Survey Scientific Investigations Report 2008-5202, 196 p.
}

